Render elements using blocks store
